No longer depending on main project pom artifact.
Although was <scope>provided</scope>, it would still cause tools to look for updates of
the dependency.  In some projects, the main project is not deployed to artifact repositores
for *-SNAPSHOT versions (for efficiency), while the books are deployed.  This would cause
build failures.

This was only used for version property substitutions, and the books and apidocs
already match the main project version by convention.

Now uses @{project.version} for the book.

For apidocs, gets the project dynamically with the new ${ao:getProjectVersion()} function,
falling-back to "¿@{project.version}?" when cannot be found (such as running in dev mode
without javadoc overlays).